Remove fileNameWithoutSuffix from public API
This had a single use in our rules and is unlikely to be needed in third-
party rule sets.
