--- detekt-psi-utils/src/test/kotlin/io/github/detekt/psi/KtFilesSpec.kt
@@ -1,46 +1,0 @@
-package io.github.detekt.psi
-
-import io.github.detekt.test.utils.internal.FakePsiFile
-import org.assertj.core.api.Assertions.assertThat
-import org.jetbrains.kotlin.com.intellij.psi.PsiFile
-import org.junit.jupiter.api.Nested
-import org.junit.jupiter.api.Test
-
-class KtFilesSpec {
-
-    @Nested
-    inner class fileNameWithoutSuffix {
-
-        @Test
-        fun `should remove kt suffix`() {
-            val filename = makeFile("C.kt").fileNameWithoutSuffix()
-            assertThat(filename).isEqualTo("C")
-        }
-
-        @Test
-        fun `should remove kts suffix`() {
-            val filename = makeFile("C.kts").fileNameWithoutSuffix()
-            assertThat(filename).isEqualTo("C")
-        }
-
-        @Test
-        fun `should not remove non kotlin suffixes`() {
-            val filename = makeFile("C.java").fileNameWithoutSuffix()
-            assertThat(filename).isEqualTo("C.java")
-        }
-
-        @Test
-        fun `should not remove common suffix`() {
-            val filename = makeFile("C.common").fileNameWithoutSuffix()
-            assertThat(filename).isEqualTo("C.common")
-        }
-
-        @Test
-        fun `should not remove common_ktx suffix`() {
-            val filename = makeFile("C.common.ktx").fileNameWithoutSuffix()
-            assertThat(filename).isEqualTo("C.common.ktx")
-        }
-    }
-
-    private fun makeFile(filename: String): PsiFile = FakePsiFile(name = filename)
-}
