+++ detekt-rules-naming/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/naming/MatchingDeclarationNameSpec.kt
@@ -1,10 +1,13 @@
 package io.gitlab.arturbosch.detekt.rules.naming
 
 import io.github.detekt.test.utils.compileContentForTest
+import io.github.detekt.test.utils.internal.FakePsiFile
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.test.TestConfig
 import io.gitlab.arturbosch.detekt.test.assertThat
 import io.gitlab.arturbosch.detekt.test.lint
+import org.assertj.core.api.Assertions
+import org.jetbrains.kotlin.com.intellij.psi.PsiFile
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 
@@ -258,4 +261,40 @@ class MatchingDeclarationNameSpec {
             assertThat(findings).hasStartSourceLocation(1, 14)
         }
     }
+
+    @Nested
+    inner class FileNameWithoutSuffix {
+
+        @Test
+        fun `should remove kt suffix`() {
+            val filename = makeFile("C.kt").fileNameWithoutSuffix()
+            Assertions.assertThat(filename).isEqualTo("C")
+        }
+
+        @Test
+        fun `should remove kts suffix`() {
+            val filename = makeFile("C.kts").fileNameWithoutSuffix()
+            Assertions.assertThat(filename).isEqualTo("C")
+        }
+
+        @Test
+        fun `should not remove non kotlin suffixes`() {
+            val filename = makeFile("C.java").fileNameWithoutSuffix()
+            Assertions.assertThat(filename).isEqualTo("C.java")
+        }
+
+        @Test
+        fun `should not remove common suffix`() {
+            val filename = makeFile("C.common").fileNameWithoutSuffix()
+            Assertions.assertThat(filename).isEqualTo("C.common")
+        }
+
+        @Test
+        fun `should not remove common_ktx suffix`() {
+            val filename = makeFile("C.common.ktx").fileNameWithoutSuffix()
+            Assertions.assertThat(filename).isEqualTo("C.common.ktx")
+        }
+    }
+
+    private fun makeFile(filename: String): PsiFile = FakePsiFile(name = filename)
 }
